ngs: Implement lock/unlock params and some module skeleton
