+++ vita3k/ngs/CMakeLists.txt
@@ -1,7 +1,15 @@
 add_library(
 	ngs
 	STATIC
+	include/ngs/modules/atrac9.h
+	include/ngs/modules/player.h
+	include/ngs/common.h
+	include/ngs/mempool.h
+	include/ngs/scheduler.h
 	include/ngs/state.h
+	include/ngs/system.h
+	src/modules/atrac9.cpp
+	src/modules/player.cpp
 	src/allocator.cpp
 	src/ngs.cpp
 	src/scheduler.cpp
