+++ vita3k/ngs/include/ngs/modules/player.h
@@ -1,0 +1,42 @@
+#pragma once
+
+#include <ngs/system.h>
+
+namespace emu::ngs::player {
+    struct BufferParameter {
+        Ptr<void> buffer;
+        std::int32_t bytes_count;
+        std::int16_t loop_count;
+        std::int16_t next_buffer_index;
+        std::int16_t samples_discard_start_off;
+        std::int16_t samples_discard_end_off;
+    };
+
+    static constexpr std::uint32_t MAX_BUFFER_PARAMS = 4;
+
+    struct Parameters {
+        emu::ngs::ModuleParameterHeader header;
+        BufferParameter buffer_params[MAX_BUFFER_PARAMS];
+        float playback_frequency;
+        float playback_scalar;
+        std::int32_t lead_in_samples;
+        std::int32_t limit_number_of_samples_played;
+        std::int32_t start_bytes;
+        std::int8_t channels;
+        std::int8_t channel_map[2];
+        std::int8_t type;
+        std::int8_t unk60;
+        std::int8_t start_buffer;
+        std::int8_t unk62;
+        std::int8_t unk63;
+    };
+
+    struct Module: public emu::ngs::VoiceDefinition {
+        explicit Module();
+        void process(Voice *voice) override;
+
+        std::size_t get_buffer_parameter_size() const override {
+            return sizeof(Parameters);
+        }
+    };
+};
