+++ vita3k/ngs/src/modules/atrac9.cpp
@@ -1,0 +1,11 @@
+#include <ngs/modules/atrac9.h>
+
+namespace emu::ngs::atrac9 {
+    Module::Module() 
+        : VoiceDefinition(emu::ngs::BUSS_ATRAC9) {
+    }
+
+    void Module::process(Voice *voice) {
+        
+    }
+};
